requirements.txt
discord.py
py-cord
selenium
beautifulsoup4